// Frontend/src/enums/CAM/CAM01/cam01.ts
export enum Cam01Status {
  All = 'All',
  Draft = 'Draft',
  InProgress = 'InProgress',
  Completed = 'Completed',
}

export enum Cam01Type {
  ChangeContractDetails = 'ChangeContractDetails',
  AppendNewPurchaseOrder = 'AppendNewPurchaseOrder',
  WaiveOrReducePenalty = 'WaiveOrReducePenalty',
  AdjustContractDuration = 'AdjustContractDuration',
}

export enum Cam01PoStep {
  PoAddendum = 'PoAddendum',
  PoSap = 'PoSap'
};